Ensure request queue is saved to persistent storage when durableCache enabled.
- providing a "pre" queue system that overcomes the limitations of saving and restoring the apollo queue

Notes:
Requests are queued during the submission request as a simple array of data from the form. Items are stored and set to the apollo client.  If the system is online, the request will complete and the saved array of requests will be cleared.  If the system is offline, the apollo queue will retain the data (UI depends on the data in the queue), clearing the queue once the system is online again.

If the system is offline and the user logs out or the browser is closed, the queued requests will be restored upon login and sent to the apollo queue for processing if possible.